requirements.txt
gitdb2==2.0.3
GitPython==2.1.9
nltk==3.3
six==1.11.0
smmap2==2.0.3
